IK8960KK-2254 : Fix rpm_stats overflow error
Propagated from:
IKQCOMSD1-305: Fix rpm_stats overflow error

Total time(uSec) value for xo_shutdown and vdd_min will be reported
incorrectly because of overflow. Casting the variable to uint64_t
to fix the overflow.

Reviewed-on: http://gerrit.pcs.mot.com/616382
SLTApproved: Slta Waiver <sltawvr@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>
Submit-Approved: Jira Key <jirakey@motorola.com>

+++ arch/arm/mach-msm/rpm_stats.c
@@ -107,7 +107,7 @@ static int msm_rpmstats_copy_stats(struct msm_rpmstats_private_data *pdata)
 						pdata->read_idx, 2);
 
 	if (record.id == ID_ACCUM_TIME_SCLK) {
-		usec = record.val * USEC_PER_SEC;
+		usec = (uint64_t)record.val * USEC_PER_SEC;
 		do_div(usec, SCLK_HZ);
 	}  else
 		usec = (unsigned long)record.val;
